requirements.txt
networkx>=2.8.0
matplotlib>=3.5.0
openai>=1.0.0
numpy>=1.22.0
spacy>=3.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
sympy>=1.9.0  # 用于数学验证
jsonlines>=2.0.0  # 用于数据集处理
requests>=2.27.0  # 用于数据下载
PyYAML>=6.0  # 用于配置管理
transformers>=4.18.0  # 用于文本相似度评估
nltk>=3.6.0  # 用于文本处理和评估
rouge_score>=0.1.2  # 用于评估指标计算